created ignore filter :: WIP

+++ app/src/main/java/com/balance/update/autobalanceupdate/domain/filter/FilterInteractor.kt
@@ -3,10 +3,14 @@ package com.balance.update.autobalanceupdate.domain.filter
 import com.balance.update.autobalanceupdate.data.db.entities.Filter
 import com.balance.update.autobalanceupdate.data.memory.DateRange
 import com.balance.update.autobalanceupdate.data.repository.FilterRepository
+import com.balance.update.autobalanceupdate.domain.CompletableInteractor
 import com.balance.update.autobalanceupdate.domain.ObservableInteractor
 import com.balance.update.autobalanceupdate.domain.SingleInteractor
+import io.reactivex.Completable
 import io.reactivex.Observable
 import io.reactivex.Single
+
+private const val FILTER_IGNORE_KEY = -1
 
 private val repository = FilterRepository()
 
@@ -34,6 +38,9 @@ class CreateFilter : SingleInteractor<Long, String>() {
 class DeleteFilter : SingleInteractor<Int, Filter>() {
 
     override fun buildCase(params: Filter): Single<Int> {
+        if (params.key == FILTER_IGNORE_KEY) {
+            return Single.error(Throwable("Ignore filter can't be removed"))
+        }
         return repository.delete(params)
     }
 
@@ -42,7 +49,18 @@ class DeleteFilter : SingleInteractor<Int, Filter>() {
 class UpdateFilter : SingleInteractor<Long, Filter>() {
 
     override fun buildCase(params: Filter): Single<Long> {
+        if (params.key == FILTER_IGNORE_KEY) {
+            return Single.error(Throwable("Ignore filter can't be updated"))
+        }
         return repository.update(params)
     }
+
+}
+
+class CreateIgnoreFilter : CompletableInteractor<Unit>() {
+    override fun buildCase(params: Unit): Completable =
+            repository.loadFilterById(FILTER_IGNORE_KEY)
+                    .ignoreElement()
+                    .onErrorResumeNext { repository.create(filterId = FILTER_IGNORE_KEY, filterName = "Ignore").ignoreElement() }
 
 }
